backend/service_provider_service/requirements.txt
# backend/booking_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
redis==5.0.8
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
sqlalchemy==2.0.19
passlib[bcrypt]
pyjwt
alembic
pydantic
pydantic-settings